flux and reactive mongo
