--- src/main/java/ru/ezikvice/springotus/homework10/controller/IndexController.java
@@ -1,20 +1,0 @@
-package ru.ezikvice.springotus.homework10.controller;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.GetMapping;
-
-@Controller
-public class IndexController {
-
-    @Autowired
-    public IndexController() {
-
-    }
-
-    @GetMapping("/")
-    public String listPage(Model model) {
-        return "index";
-    }
-}
