--- src/main/java/ru/ezikvice/springotus/homework10/dao/BookRepository.java
@@ -1,11 +1,0 @@
-package ru.ezikvice.springotus.homework10.dao;
-
-import org.springframework.data.mongodb.repository.MongoRepository;
-import ru.ezikvice.springotus.homework10.domain.Book;
-
-public interface BookRepository extends MongoRepository<Book, Long> {
-
-    Book save(Book book);
-
-    Book findById(String id);
-}
