--- src/main/java/ru/ezikvice/springotus/homework10/domain/Author.java
@@ -1,26 +1,0 @@
-package ru.ezikvice.springotus.homework10.domain;
-
-import lombok.Data;
-import lombok.NoArgsConstructor;
-import org.springframework.data.annotation.Id;
-import org.springframework.data.mongodb.core.mapping.Document;
-
-@Document
-@Data
-@NoArgsConstructor
-public class Author {
-
-    @Id
-    private String id;
-
-    private String name;
-
-    public Author(String name) {
-        this.name = name;
-    }
-
-    public Author(String id, String name) {
-        this.id = id;
-        this.name = name;
-    }
-}
