--- src/main/java/ru/ezikvice/springotus/homework10/service/AuthorServiceImpl.java
@@ -1,50 +1,0 @@
-package ru.ezikvice.springotus.homework10.service;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-import ru.ezikvice.springotus.homework10.dao.AuthorRepository;
-import ru.ezikvice.springotus.homework10.domain.Author;
-
-import java.util.List;
-import java.util.Set;
-
-@Service
-public class AuthorServiceImpl implements AuthorService {
-
-    private final AuthorRepository authorRepository;
-
-    @Autowired
-    public AuthorServiceImpl(AuthorRepository authorRepository) {
-        this.authorRepository = authorRepository;
-    }
-
-    @Override
-    public long count() {
-        return authorRepository.count();
-    }
-
-    @Override
-    public Author save(Author author) {
-        return authorRepository.save(author);
-    }
-
-    @Override
-    public void delete(Author author) {
-        authorRepository.delete(author);
-    }
-
-    @Override
-    public Author findById(String id) {
-        return authorRepository.findById(id);
-    }
-
-    @Override
-    public List<Author> findAll() {
-        return authorRepository.findAll();
-    }
-
-    @Override
-    public Set<Author> findByName(String name) {
-        return authorRepository.findByName(name);
-    }
-}
