+++ src/main/java/ru/ezikvice/springotus/homework11/Homework11Application.java
@@ -1,0 +1,19 @@
+package ru.ezikvice.springotus.homework11;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class Homework11Application {
+
+    public static void main(String[] args) {
+        SpringApplication.run(Homework11Application.class);
+    }
+
+//    @Bean
+//    public RouterFunction<ServerResponse> monoRouterFunction() {
+//        HandlerFunction<ServerResponse> hello = request ->
+//                ServerResponse.ok().body(fromPublisher(Flux.range(1, 10), Integer.class));
+//        return route(GET("/hello"), hello);
+//    }
+}
