+++ src/main/java/ru/ezikvice/springotus/homework11/config/AppConfig._java
@@ -1,0 +1,20 @@
+package ru.ezikvice.springotus.homework11.config;
+
+import com.mongodb.MongoClient;
+import org.springframework.context.annotation.Bean;
+import org.springframework.data.mongodb.config.AbstractReactiveMongoConfiguration;
+import org.springframework.data.mongodb.repository.config.EnableReactiveMongoRepositories;
+
+@EnableReactiveMongoRepositories
+public class AppConfig extends AbstractReactiveMongoConfiguration {
+
+    @Bean
+    public MongoClient mongoClient() {
+        return MongoClients.create();
+    }
+
+    @Override
+    protected String getDatabaseName() {
+        return "reactive";
+    }
+}
