+++ src/main/java/ru/ezikvice/springotus/homework11/service/AuthorServiceImpl.java
@@ -1,0 +1,52 @@
+package ru.ezikvice.springotus.homework11.service;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+import reactor.core.publisher.Flux;
+import reactor.core.publisher.Mono;
+import ru.ezikvice.springotus.homework11.dao.AuthorRepository;
+import ru.ezikvice.springotus.homework11.domain.Author;
+
+import java.util.List;
+import java.util.Set;
+
+@Service
+public class AuthorServiceImpl implements AuthorService {
+
+    private final AuthorRepository authorRepository;
+
+    @Autowired
+    public AuthorServiceImpl(AuthorRepository authorRepository) {
+        this.authorRepository = authorRepository;
+    }
+
+    @Override
+    public Mono<Long> count() {
+        return authorRepository.count();
+    }
+
+    @Override
+    public Mono<Author> save(Author author) {
+        return authorRepository.save(author);
+    }
+
+    @Override
+    public void delete(Author author) {
+        authorRepository.delete(author);
+    }
+
+    @Override
+    public Mono<Author> findById(String id) {
+        return authorRepository.findById(id);
+    }
+
+    @Override
+    public Flux<Author> findAll() {
+        return authorRepository.findAll();
+    }
+
+    @Override
+    public Flux<Set<Author>> findByName(String name) {
+        return authorRepository.findByName(name);
+    }
+}
