--- src/test/resources/application.yml
@@ -1,13 +1,0 @@
-# Thanks to Mkyong for mysql and hikari settings
-# http://www.mkyong.com/spring-boot/spring-boot-jdbc-mysql-hikaricp-example/
-
-#disbale Spring banner
-# By quoting the value, you disable yaml's automatic-cast-to-boolean feature.
-# https://github.com/spring-projects/spring-boot/issues/4600#issuecomment-219687848
-spring:
-  main:
-    banner-mode: 'OFF'
-  shell:
-    interactive:
-      enabled: false
-
